WIP: Some meaningful output

+++ src/Tools/dotnet-trace-hosted/Controllers/MonitorsController.cs
@@ -17,7 +17,7 @@ namespace HostedTrace
         [HttpPost]
         public async Task<ActionResult> StartAsync([FromBody]MonitorRequest requestParam)
         {
-            int sessionId = await _monitorService.StartMonitorAsync(requestParam.ProcessId).ConfigureAwait(false);
+            ulong sessionId = await _monitorService.StartMonitorAsync(requestParam.ProcessId).ConfigureAwait(false);
             return Ok(sessionId);
         }
     }
